Check with the database the registration status.
At the first login, the custom claims will set 'hasCompletedRegistration' to false, and the token can't be updated afterwards. This means that when the client logs in for the first time and registers afterwards, this endpoint would return that the user is not actually registered. Querying the database fixes this, asthe current registration status is retrieved this way.